+++ CHANGELOG.md
@@ -10,6 +10,8 @@
 - Updated "Pitch Black" achievement requirement to 6 Inklings in a row (was 7)
 - Enhanced collection tracker to support item variant multipliers (e.g., Enchanted Ink Sac now counts as 160)
   - I tested ink collection and skyblock didn't count it at all though, like i think its bugged
+- Updated Umberella display and alerts to work like the Flare display, using a 30-block range check
+
 ### Fixes
 - Fixed typo in plhleg emoji
 - Fixed Frog Blessing Display not working